best shell ever

+++ src/launch/job_process_status.c
@@ -18,7 +18,8 @@ static int		wait_for_job_ext(t_job_pid *job_pids, int id)
 	{
 		job_pids->wait_pid = waitpid(-g_sh->jobs[id]->pgid,
 				&job_pids->status, WUNTRACED);
-		kill(g_sh->jobs[id]->root->pid, SIGQUIT);
+		if (!(g_sh->jobs[id]->root->type))
+            kill(g_sh->jobs[id]->root->pid, SIGQUIT);
 		job_pids->wait_count++;
 		if (WIFEXITED(job_pids->status))
 		{
